+++ core/src/components.d.ts
@@ -307,6 +307,10 @@ export namespace Components {
     }
     interface CatFormGroup {
         /**
+          * Whether the label is on top or left.
+         */
+        "horizontal": boolean;
+        /**
           * The space allocated to the input label
          */
         "labelSize"?: string;
@@ -478,6 +482,10 @@ export namespace Components {
           * The unique identifier for a referenced input.
          */
         "for"?: string;
+        /**
+          * Whether the label is on top or left.
+         */
+        "horizontal": boolean;
         /**
           * A value is required or must be check for the form to be submittable.
          */
@@ -700,6 +708,10 @@ export namespace Components {
           * Optional hint text(s) to be displayed with the select.
          */
         "hint"?: string | string[];
+        /**
+          * Whether the label is on top or left.
+         */
+        "horizontal": boolean;
         /**
           * A unique identifier for the input.
          */
@@ -1627,6 +1639,10 @@ declare namespace LocalJSX {
     }
     interface CatFormGroup {
         /**
+          * Whether the label is on top or left.
+         */
+        "horizontal"?: boolean;
+        /**
           * The space allocated to the input label
          */
         "labelSize"?: string;
@@ -1793,6 +1809,10 @@ declare namespace LocalJSX {
           * The unique identifier for a referenced input.
          */
         "for"?: string;
+        /**
+          * Whether the label is on top or left.
+         */
+        "horizontal"?: boolean;
         /**
           * A value is required or must be check for the form to be submittable.
          */
@@ -2033,6 +2053,10 @@ declare namespace LocalJSX {
           * Optional hint text(s) to be displayed with the select.
          */
         "hint"?: string | string[];
+        /**
+          * Whether the label is on top or left.
+         */
+        "horizontal"?: boolean;
         /**
           * A unique identifier for the input.
          */