+++ core/src/components/cat-textarea/cat-textarea.tsx
@@ -205,7 +205,7 @@ export class CatTextarea {
         : value === true
         ? {}
         : value || undefined;
-      this.showErrorsAfterTimeout();
+      this.showErrorsIfTimeout() || this.showErrorsIfNoFocus();
     }
   }
 
@@ -306,7 +306,7 @@ export class CatTextarea {
   private onInput(event: InputEvent) {
     this.value = this.textarea.value;
     this.catChange.emit(event);
-    this.showErrorsAfterTimeout();
+    this.showErrorsIfTimeout();
   }
 
   private onFocus(event: FocusEvent) {
@@ -325,11 +325,20 @@ export class CatTextarea {
   }
 
   private errorUpdateTimeoutId?: number;
-  private showErrorsAfterTimeout() {
+  private showErrorsIfTimeout() {
     const errorUpdate = coerceNumber(this.errorUpdate, null);
     if (errorUpdate !== null) {
       typeof this.errorUpdateTimeoutId === 'number' && window.clearTimeout(this.errorUpdateTimeoutId);
       this.errorUpdateTimeoutId = window.setTimeout(() => this.showErrors(), errorUpdate);
+      return true;
+    }
+    return false;
+  }
+
+  private showErrorsIfNoFocus() {
+    const hasFocus = document.activeElement === this.hostElement || document.activeElement === this.textarea;
+    if (!hasFocus) {
+      this.showErrors();
     }
   }
 }